+++ app/src/main/java/garmingoproremote/android/app/ServiceManager.java
@@ -1,4 +1,4 @@
-package com.example.garmingopromobile;
+package garmingoproremote.android.app;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
