+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <activity
         android:id="@+id/mainActivity"
-        android:name="com.example.garmingopromobile.MainActivity"
+        android:name="garmingoproremote.android.app.MainActivity"
         android:label="activity_main"
         tools:layout="@layout/activity_main" />
 </navigation>
