+++ app/src/test/java/garmingoproremote/android/app/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.garmingopromobile;
+package garmingoproremote.android.app;
 
 import org.junit.Test;
